Update Embedded Book for mdbook 0.2.0
mdbook 0.2.0's has a breaking change such that links are relative to the file they are in instead of the top level of the book. Luckily, this is your only link like that. So once you accept this PR you can build with mdbook 0.2.0 which is what we're trying to switch to for the the bookshelf.